pyproject.toml
[project]
name = "backflip-sdr-agent-teams"
version = "0.1.0"
description = "Backflip Media SDR Agent pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core framework
    "google-adk>=1.0.0",
    "google-cloud-aiplatform>=1.60.0",
    "litellm>=1.40.0",
    # Lead discovery
    "exa-py>=1.0.0",
    # Contact enrichment (Hunter.io via REST)
    "requests>=2.31.0",
    # Google Calendar
    "google-auth>=2.28.0",
    "google-api-python-client>=2.120.0",
    # Voice
    "elevenlabs>=1.0.0",
    # Data validation
    "pydantic>=2.0.0",
    # Utilities
    "python-dotenv>=1.0.0",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.29",
    # Observability
    "langfuse>=2.0",
    "python-dateutil>=2.8",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
